GO-5419: Refactor and fix filesync bugs

+++ core/files/filesync/queue.go
@@ -17,7 +17,7 @@ type QueueItem struct {
 	SpaceId     string
 	ObjectId    string
 	FileId      domain.FileId
-	Timestamp   int64
+	Timestamp   float64
 	AddedByUser bool
 	Imported    bool
